+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Conflict manager can now re-run the patching step for quick changes
 - Meshes with no plugin uses and not flagged for patching will not be loaded during patching step
 - Fixed sometimes ambiguous matching priority when multiple matches of different shader types come from the same mod
+- Removed --highmem CLI argument
 
 ## [1.1.4] - 2026-06-24